+++ packages/backend/src/services/ProjectService/ProjectService.ts
@@ -27,7 +27,6 @@ import {
     DashboardBasicDetails,
     DashboardFilters,
     DateGranularity,
-    dateGranularityToTimeFrameMap,
     DbtExposure,
     DbtExposureType,
     DbtProjectType,
@@ -43,7 +42,6 @@ import {
     FilterOperator,
     findFieldByIdInExplore,
     ForbiddenError,
-    formatDate,
     formatRows,
     getAggregatedField,
     getDashboardFilterRulesForTables,
@@ -51,7 +49,6 @@ import {
     getDimensions,
     getFieldQuoteChar,
     getFields,
-    getFirstDayOfQuarter,
     getIntrinsicUserAttributes,
     getItemId,
     getMetrics,
@@ -64,7 +61,6 @@ import {
     isExploreError,
     isFilterableDimension,
     isFilterRule,
-    isOrFilterGroup,
     isUserWithOrg,
     ItemsMap,
     Job,
@@ -101,9 +97,7 @@ import {
     SummaryExplore,
     TablesConfiguration,
     TableSelectionType,
-    TimeFrames,
     UnexpectedServerError,
-    updateFilterValueInFilters,
     UpdateMetadata,
     UpdateProject,
     UpdateProjectMember,
@@ -1053,56 +1047,12 @@ export class ProjectService extends BaseService {
         return { adapter, sshTunnel };
     }
 
-    static updateMetricQueryWithGranularity(
-        updatedDimensions: string[],
-        metricQuery: MetricQuery,
-        granularity?: DateGranularity,
-    ): MetricQuery {
-        if (granularity && updatedDimensions.length > 0) {
-            const dimensionFilters = metricQuery.filters.dimensions;
-            if (!dimensionFilters) return metricQuery;
-
-            // These filters are updated by reference
-            updateFilterValueInFilters(
-                dimensionFilters,
-                (fieldId: string, filterValues: any[] | undefined) => {
-                    if (
-                        updatedDimensions.includes(fieldId) &&
-                        filterValues?.[0]
-                    ) {
-                        // Apply date granularity to filter value
-                        const filterValue = filterValues[0];
-                        const timeFrame =
-                            dateGranularityToTimeFrameMap[granularity];
-
-                        if (timeFrame === TimeFrames.QUARTER) {
-                            // Get first day of quarter using dayjs
-                            const newDate = getFirstDayOfQuarter(filterValue);
-                            return [newDate];
-                        }
-                        const newDate = formatDate(filterValue, timeFrame);
-                        return [newDate];
-                    }
-                    return filterValues;
-                },
-            );
-            return {
-                ...metricQuery,
-                filters: {
-                    ...metricQuery.filters,
-                    dimensions: dimensionFilters,
-                },
-            };
-        }
-        return metricQuery;
-    }
-
     static updateExploreWithGranularity(
         explore: Explore,
         metricQuery: MetricQuery,
         warehouseClient: WarehouseClient,
         granularity?: DateGranularity,
-    ): { explore: Explore; updatedDimensions: string[] } {
+    ): Explore {
         if (granularity) {
             const timeDimensionsMap: Record<string, CompiledDimension> =
                 Object.values(explore.tables).reduce<
@@ -1141,17 +1091,13 @@ export class ProjectService extends BaseService {
                         warehouseClient,
                         granularity,
                     );
-                const updatedExplore = replaceDimensionInExplore(
+                return replaceDimensionInExplore(
                     explore,
                     dimWithGranularityOverride,
                 );
-                return {
-                    explore: updatedExplore,
-                    updatedDimensions: [getItemId(dimToOverride)],
-                };
             }
         }
-        return { explore, updatedDimensions: [] };
+        return explore;
     }
 
     static async _compileQuery(
@@ -1163,23 +1109,16 @@ export class ProjectService extends BaseService {
         timezone: string,
         granularity?: DateGranularity,
     ): Promise<CompiledQuery> {
-        const { explore: exploreWithOverride, updatedDimensions } =
-            ProjectService.updateExploreWithGranularity(
-                explore,
-                metricQuery,
-                warehouseClient,
-                granularity,
-            );
+        const exploreWithOverride = ProjectService.updateExploreWithGranularity(
+            explore,
+            metricQuery,
+            warehouseClient,
+            granularity,
+        );
 
-        const metricWithOverride =
-            ProjectService.updateMetricQueryWithGranularity(
-                updatedDimensions,
-                metricQuery,
-                granularity,
-            );
         const compiledMetricQuery = compileMetricQuery({
             explore: exploreWithOverride,
-            metricQuery: metricWithOverride,
+            metricQuery,
             warehouseClient,
         });